Moving to conf system - ColorMixin 02
