+++ src/base/i-color-mixin-conf.ts
@@ -1,6 +1,7 @@
 import {IBaseMixinConf} from './i-base-mixin-conf';
-import {BaseAccessor} from '../core/types';
+import {BaseAccessor, ColorAccessor} from '../core/types';
 
 export interface IColorMixinConf extends IBaseMixinConf {
+    colorAccessor?: ColorAccessor;
     colorCalculator?: BaseAccessor<string>;
 }
