+++ src/charts/sunburst-chart.ts
@@ -85,7 +85,7 @@ export class SunburstChart extends ColorMixin(BaseMixin) {
         this._minAngleForLabel = DEFAULT_MIN_ANGLE_FOR_LABEL;
         this._externalLabelRadius = undefined;
 
-        this.colorAccessor(d => this._conf.keyAccessor(d));
+        this._conf.colorAccessor = d => this._conf.keyAccessor(d);
 
         // override cap mixin // TODO: not needed, does not mix CapMixin any longer
         this._conf.ordering = d => d.key;
